+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## In development
+
+## 0.5.0 (Dec 26, 2014)
+
+ - **(POTENTIALLY BREAKING CHANGE)** Make `recalc` on `Polygon` more memory efficient. It no longer does any allocations. The `calcPoints`,`edges` and `normals` vector arrays are reused and only created in `setPoints` when the number of new points is different than the current ones. (Fixes #15)
+   - `points`, `angle` and `offset` can no longer be manually changed. The `setPoints`, `setAngle`, and `setOffset` methods **must** be used.
+   - As a result of this, the `recalc` method is no longer part of the API.
+ - Add `getAABB` to `Polygon` and `Circle` that calculate Axis-Aligned Bounding Boxes - thanks [TuurDutoit](https://github.com/TuurDutoit)! (Fixes #17)
+
+## 0.4.1 (Mar 23, 2014)
+
+ - Fix missing `T_VECTORS.push()` - thanks [shakiba](https://github.com/shakiba)! (Fixes #8)
+ - Add `package.json` - released as `npm` module (Fixes #11, Fixes #12)
+
 ## 0.4 (Mar 2, 2014)
 
  - Add `clone` method to `Vector` that returns a new vector with the same coordinates.
@@ -25,4 +39,4 @@
 
 ## 0.1
 
-Initial release
+ - Initial release